+++ code/mesh/src/mesh_manipulations.cpp
@@ -33,7 +33,7 @@ Eigen::Vector3d barycenter(const Matrix3x& p)
     {
         points.push_back(p.col(j));
     }
-    return sum::pairwise(points)/points.size();
+    return sum::pairwise(points)/double(points.size());
 }
 
 Eigen::Vector3d unit_normal(const Matrix3x& points)
